+++ android/src/main/java/com/smileidentity/react/SmileIdPackage.kt
@@ -11,12 +11,14 @@ import com.smileidentity.react.viewmanagers.SmileIDConsentViewManager
 import com.smileidentity.react.viewmanagers.SmileIDDocumentVerificationViewManager
 import com.smileidentity.react.viewmanagers.SmileIDEnhancedDocumentVerificationViewManager
 import com.smileidentity.react.viewmanagers.SmileIDSmartSelfieAuthenticationViewManager
+import com.smileidentity.react.viewmanagers.SmileIDSmartSelfieCaptureViewManager
 import com.smileidentity.react.viewmanagers.SmileIDSmartSelfieEnrollmentViewManager
 
 class SmileIdPackage : TurboReactPackage() {
 
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> =
     listOf(
+      SmileIDSmartSelfieCaptureViewManager(reactContext),
       SmileIDSmartSelfieEnrollmentViewManager(reactContext),
       SmileIDSmartSelfieAuthenticationViewManager(reactContext),
       SmileIDDocumentVerificationViewManager(reactContext),