+++ example/src/SmileIDCaptureScreen.tsx
@@ -13,6 +13,7 @@ import {
   SmileID,
   IdInfo,
   JobStatusRequest,
+  SmileIDSmartSelfieCaptureView,
 } from '@smile_identity/react-native';
 
 import type {
@@ -128,6 +129,26 @@ export const SmileIDCaptureScreen: React.FC<SmileIDCaptureScreenProps> = ({
 
   return (
     <View style={styles.container}>
+      {title === 'SmartSelfie Capture' && (
+        // @ts-ignore - this is a known issue with the type definitions
+        <SmileIDSmartSelfieCaptureView
+          {...product}
+          style={[styles.smileView, { backgroundColor: '#fff' }]}
+          onResult={(event) => {
+            if (event.nativeEvent.error) {
+              handleErrorResponse(event.nativeEvent.error);
+              return;
+            }
+            handleResponse(
+              JobType.SmartSelfieEnrollment,
+              'pollSmartSelfieJobStatus',
+              event.nativeEvent.result,
+              product.userId,
+              product.jobId
+            );
+          }}
+        />
+      )}
       {title === 'SmartSelfie Enrollment' && (
         // @ts-ignore - this is a known issue with the type definitions
         <SmileIDSmartSelfieEnrollmentView